requirements.txt
customtkinter>=5.2.0
opencv-python>=4.8.0
pyzbar>=0.1.9
pandas>=2.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
tkcalendar>=1.6.1
qrcode>=7.4.2
reportlab>=4.2.0
